#  Multiple optimizations to the viewer function 