--- a/functions_f_stringtable/functions/fn_stringtable_viewer.sqf
+++ b/functions_f_stringtable/functions/fn_stringtable_viewer.sqf
@@ -22,13 +22,9 @@ params
 #define BUSY_BACKGROUND (DISPLAY displayCtrl 55000)
 #define BUSY_BACKGROUND_PRELOADING (DISPLAY displayCtrl 56000)
 
-switch (tolower _mode) do
+switch _mode do
 {
-        case ("exit"):
-        {
-                DISPLAY closeDisplay 2;
-        };
-        case ("onload"):
+        case "onload":
         {
                 _params params [["_display",displayNull,[displayNull]]];
                 uiNamespace setVariable ["asaayu_stringtable_display",_display];
@@ -37,51 +33,26 @@ switch (tolower _mode) do
                 missionNamespace setVariable ["stringtable_viewer_origin_index",0];
                 missionNamespace setVariable ["stringtable_viewer_language",language];
 
-                {
-                        if ((tolower configName _x) isEqualTo (tolower language)) then
-                        {
-                                missionNamespace setVariable ["stringtable_viewer_language_index",_foreachindex];
-                        };
-                } foreach ("true" configClasses (configfile >> "CfgLanguages"));
+                private _languages = ("true" configClasses (configfile >> "CfgLanguages")) apply {configname _x};
+                missionNamespace setVariable ["stringtable_viewer_language_index",_languages find language];
 
                 ["preload"] spawn STRINGTABLE_fnc_stringtable_viewer;
         };
-        case ("onunload"):
+        case "preload":
         {
-                _params params [["_display",displayNull,[displayNull]]];
-                _display closeDisplay 2;
-                uiNamespace setVariable ["asaayu_stringtable_display",displayNull];
-        };
-        case ("preload"):
-        {
-                private _check = (uiNamespace getVariable ["stringtable_viewer_allow_preload",true]);
-                if (_check) then
+                if (uiNamespace getVariable ["stringtable_viewer_allow_preload",true]) then
                 {
                         private _diag_ticktime = diag_ticktime;
 
                         startLoadingScreen [""];
-                        private _languages = ("true" configClasses (configfile >> "CfgLanguages"));
-                        private _origins = ("true" configClasses (configfile >> "cfgstringtables"));
-                        private _master = [];
-                        {
-                                private _config = _x;
-                                private _configname = configName _config;
-                                private _keys = ("true" configClasses (_config));
-                                private _data = [];
-                                {
-                                        private _key_text = [];
-                                        private _config = _x;
-                                        private _configname = configName _x;
-                                        {
-                                                // Need to parseText because some contin html entites :|
-                                                private _text = str (parseText (getText (_config >> (configName _x))));
-                                                _key_text pushBack _text
-                                        } foreach _languages;
-                                        _data pushBack [_configname,_key_text];
-                                } foreach _keys;
-
-                                _master pushBack [_configname, _data];
-                        } foreach _origins;
+                        private _languages = ("true" configClasses (configfile >> "CfgLanguages")) apply {configname _x};
+                        private _master = ("true" configClasses (configfile >> "cfgstringtables")) apply {
+                                [configname _x,("true" configClasses _x) apply {
+                                        private _key = _x;
+                                       [configname _x,_languages apply {str parseText getText(_key >> _x)}]
+                                }]
+                        };
+                        
                         uiNamespace setVariable ["stringtable_viewer_allow_preload",false];
                         uiNamespace setVariable ["stringtable_viewer_data",_master];
                         endLoadingScreen;
@@ -95,10 +66,14 @@ switch (tolower _mode) do
                 ["initgui"] call STRINGTABLE_fnc_stringtable_viewer;
                 ["eventhandlers"] call STRINGTABLE_fnc_stringtable_viewer;
                 ["loadstringtable"] spawn STRINGTABLE_fnc_stringtable_viewer;
+
+                SEARCH_EDIT ctrlSetText "";
+                SEARCH_EDIT ctrlSetTooltip "";
+                SEARCH_EDIT ctrlCommit 0;
         };
-        case ("initgui"):
+        case "initgui":
         {
-                private _languages = ("true" configClasses (configfile >> "CfgLanguages"));
+                private _languages = "true" configClasses (configfile >> "CfgLanguages");
                 {
                         private _name = getText (_x >> "name");
                         private _configname = configName _x;
@@ -109,7 +84,7 @@ switch (tolower _mode) do
                                 LANGUAGE_COMBO lbSetCurSel _index;
                         };
                 } foreach _languages;
-                private _origins = ("true" configClasses (configfile >> "cfgstringtables"));
+                private _origins = "true" configClasses (configfile >> "cfgstringtables");
                 {
                         private _name = getText (_x >> "displayName");
                         private _configname = configName _x;
@@ -120,83 +95,76 @@ switch (tolower _mode) do
 
                 SEARCH_EDIT ctrlSetText "";
 
-                { _x ctrlCommit 0 } foreach [LIST, SEARCH_EDIT, SEARCH_BUTTON, COPY_BUTTON, EXPORT_BUTTON, ORIGIN_COMBO, LANGUAGE_COMBO];
+                { _x ctrlCommit 0 } count [LIST, SEARCH_EDIT, SEARCH_BUTTON, COPY_BUTTON, EXPORT_BUTTON, ORIGIN_COMBO, LANGUAGE_COMBO];
         };
-        case ("eventhandlers"):
+        case "eventhandlers":
         {
-                STRINGTABLE_KEYDOWN_EVH = DISPLAY displayAddEventHandler ["KeyDown",{ ["keydown",_this] call STRINGTABLE_fnc_stringtable_viewer }];
-                ORIGIN_COMBO ctrlAddEventHandler ["LBSelChanged",{ ["lbselchanged",[0,(_this#1)]] call STRINGTABLE_fnc_stringtable_viewer }];
-                LANGUAGE_COMBO ctrlAddEventHandler ["LBSelChanged",{ ["lbselchanged",[1,(_this#1)]] call STRINGTABLE_fnc_stringtable_viewer }];
-                SEARCH_BUTTON ctrlAddEventHandler ["ButtonClick",{ ["search",[tolower (ctrlText SEARCH_EDIT)]] spawn STRINGTABLE_fnc_stringtable_viewer }];
+                DISPLAY displayAddEventHandler ["KeyDown",{ ["keydown",_this] call STRINGTABLE_fnc_stringtable_viewer }];
+                ORIGIN_COMBO ctrlAddEventHandler ["LBSelChanged",{ ["lbselchanged",[0,_this#1]] call STRINGTABLE_fnc_stringtable_viewer }];
+                LANGUAGE_COMBO ctrlAddEventHandler ["LBSelChanged",{ ["lbselchanged",[1,_this#1]] call STRINGTABLE_fnc_stringtable_viewer }];
+                SEARCH_BUTTON ctrlAddEventHandler ["ButtonClick",{ ["search",[tolower ctrlText SEARCH_EDIT]] spawn STRINGTABLE_fnc_stringtable_viewer }];
                 COPY_BUTTON ctrlAddEventHandler ["ButtonClick",{ ["KeyDown",[nil,DIK_C,false,true]] call STRINGTABLE_fnc_stringtable_viewer }];
                 EXPORT_BUTTON ctrlAddEventHandler ["ButtonClick",{ ["KeyDown",[nil,DIK_X,false,true]] call STRINGTABLE_fnc_stringtable_viewer }];
         };
-        case ("keydown"):
+        case "keydown":
         {
                 _params params ["_control", "_keydown_key", "_shift", "_ctrl", "_alt"];
-                private _key = (["getkey"] call STRINGTABLE_fnc_stringtable_viewer);
-                private _return = false;
-                switch (true) do
+                private _key = LIST lnbText [lbCurSel LIST,0];
+                switch true do
                 {
                         // Enter, Numpad Enter
                         case (_keydown_key in [DIK_RETURN,DIK_NUMPADENTER]):
                         {
-                                ["search",[tolower (ctrlText SEARCH_EDIT)]] spawn STRINGTABLE_fnc_stringtable_viewer;
-                                _return = true;
+                                ["search",[tolower ctrlText SEARCH_EDIT]] spawn STRINGTABLE_fnc_stringtable_viewer;
+                                true
                         };
                         // Ctrl-C
-                        case (_ctrl && {_keydown_key in [DIK_C] && {!(_key isEqualTo "")}}):
+                        case (_ctrl && {_keydown_key == DIK_C && {_key != ""}}):
                         {
                                 playSound "RscDisplayCurator_ping01";
                                 copyToClipboard format['%1',_key];
-                                _return = true;
+                                true
                         };
                         // Ctrl-X
-                        case (_ctrl && {_keydown_key in [DIK_X] && {!(_key isEqualTo "")}}):
+                        case (_ctrl && {_keydown_key == DIK_X && {_key != ""}}):
                         {
 
                                 playSound "RscDisplayCurator_ping06";
                                 copyToClipboard format['localize "%1"',_key];
-                                _return = true;
+                                true
                         };
+                        default {false};
                 };
-                _return
         };
-        case ("lbselchanged"):
+        case "lbselchanged":
         {
                 _params params ["_ctrl_index","_index"];
-                private _ctrl = ([ORIGIN_COMBO, LANGUAGE_COMBO] select _ctrl_index);
+                private _ctrl = [ORIGIN_COMBO, LANGUAGE_COMBO] select _ctrl_index;
                 private _old_variables = [stringtable_viewer_origin,stringtable_viewer_language];
-                switch (_ctrl_index) do
+                switch _ctrl_index do
                 {
                         // Origin
                         case 0:
                         {
-                                stringtable_viewer_origin = (_ctrl lbData _index);
+                                stringtable_viewer_origin = _ctrl lbData _index;
                                 stringtable_viewer_origin_index = _index;
                         };
                         // Language
                         case 1:
                         {
-                                stringtable_viewer_language = (_ctrl lbData _index);
+                                stringtable_viewer_language = _ctrl lbData _index;
                                 stringtable_viewer_language_index = _index;
                         };
                 };
                 if !(_old_variables isEqualTo [stringtable_viewer_origin,stringtable_viewer_language]) then
                 {
-                        ["search",[tolower (ctrlText SEARCH_EDIT)]] spawn STRINGTABLE_fnc_stringtable_viewer;
+                        ["search",[tolower ctrlText SEARCH_EDIT]] spawn STRINGTABLE_fnc_stringtable_viewer;
                 };
         };
-        case ("getkey"):
+        case "loadstringtable":
         {
-                private _index = lbCurSel LIST;
-                if (_index > -1) then
-                {
-                        (LIST lnbText [_index,0])
-                } else { "" };
-        };
-        case ("loadstringtable"):
-        {
+                _params params [["_search_term",""]];
+
                 lnbClear LIST;
 
                 private _diag_ticktime = diag_ticktime;
@@ -205,64 +173,23 @@ switch (tolower _mode) do
                 if (count _keys < 0) exitWith { systemchat "Error: View data < 0"; };
                 {
                         _x params ["_key","_text_list"];
-                        private _text = (_text_list#stringtable_viewer_language_index);
-                        private _row = LIST lnbAddRow [_key, _text, str 1];
-                        LIST lnbSetTooltip [[_row,0], _text];
+                        private _text = _text_list#stringtable_viewer_language_index;
+                        if (_search_term isEqualTo "" || {_search_term in toLower _key || _search_term in toLower _text}) then {
+                                private _row = LIST lnbAddRow [_key, _text, str 1];
+                                LIST lnbSetTooltip [[_row,0], _text];
+                        };
                 } foreach ((_keys#stringtable_viewer_origin_index)#1);
 
                 private _time = diag_ticktime - _diag_ticktime;
                 diag_log format[localize "STR_STRINGTABLE_INFO_DIAG_LOG",stringtable_viewer_origin,stringtable_viewer_language,str _time];
-
-                SEARCH_EDIT ctrlSetText "";
-                SEARCH_EDIT ctrlSetTooltip "";
-                SEARCH_EDIT ctrlCommit 0;
         };
-        case ("search"):
+        case "search":
         {
-                _params params ["_search_term"];
-
                 BUSY_BACKGROUND ctrlShow true;
                 ctrlSetFocus BUSY_BACKGROUND;
                 BUSY_BACKGROUND ctrlCommit 0;
 
-                private _diag_ticktime = diag_ticktime;
-
-                lnbClear LIST;
-                private _keys = uiNamespace getVariable ["stringtable_viewer_data",[]];
-                {
-                        _x params ["_key","_text_list"];
-                        private _original_key = _key;
-                        private _original_text = (_text_list#stringtable_viewer_language_index);
-                        private _text = toLower _original_text;
-                        private _key = toLower _original_key;
-                        private _value = switch (true) do
-                        {
-                                case (_search_term isEqualTo ""):
-                                {
-                                        1
-                                };
-                                case ((_key find _search_term) > -1):
-                                {
-                                        1
-                                };
-                                case ((_text find _search_term) > -1):
-                                {
-                                        1
-                                };
-                                default
-                                {
-                                        0
-                                };
-                        };
-                        if (_value > 0) then
-                        {
-                                private _row = LIST lnbAddRow [_original_key, _original_text, str 1];
-                                LIST lnbSetTooltip [[_row,0], _original_text];
-                        };
-                } foreach ((_keys#stringtable_viewer_origin_index)#1);
-
-                private _time = diag_ticktime - _diag_ticktime;
-                diag_log format[localize "STR_STRINGTABLE_INFO_DIAG_LOG",stringtable_viewer_origin,stringtable_viewer_language,str _time];
+                ["loadstringtable",_params] call stringtable_fnc_stringtable_viewer;
 
                 BUSY_BACKGROUND ctrlShow false;
                 BUSY_BACKGROUND ctrlCommit 0;