Improved performance of some tasks:
- Preloading data (after config rapify solution is implemented): 2856 ms -> 2393 ms
- Setting `stringtable_viewer_language_index`: 0.0481 ms -> 0.0259 ms
- Search filter performance: 0.0121 ms -> 0.0094 ms

Improved formatting of UI configs for better readability
Reduced codebase by ~70 lines to avoid unnecessary code duplicates and over-complication of tasks